refactor: finish alignment pipeline decomposition
